Update README.md
In the Cape Fear Estuary in North Carolina there is a substrate that is influencing the aerial biomass production. The data set has fourteen predictors that characterize the substrates physicochemical properties that are in the soil.Let the variable BIO represent the biomass production that influences by the presence of the substrate. The physicochemical properties include : H2S, Salinity, EH_7,pH, BUF, P,K, Ca, Mg, Na, Mn, Zn, Cu, and NH4. This project’s objective is to identify the important physicochemical properties of the substrate that influence the aerial biomass production in the Cape Fear Estuary.